Refactor validate functions into item and batch subpackages
