+++ src/anomalib/dataclasses/validate/tensor/__init__.py
@@ -3,26 +3,25 @@
 # Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .common import (
+from .item import (
     validate_anomaly_map,
     validate_dimensions,
+    validate_image,
     validate_label,
     validate_mask,
     validate_pred_label,
     validate_pred_mask,
     validate_pred_score,
 )
-from .image import validate_image
 
 __all__ = [
-    # Common torch validations
+    # Common torch data item validations
     "validate_anomaly_map",
     "validate_dimensions",
+    "validate_image",
     "validate_label",
     "validate_mask",
     "validate_pred_label",
     "validate_pred_mask",
     "validate_pred_score",
-    # Image specific torch validations
-    "validate_image",
 ]
