+++ src/anomalib/dataclasses/validate/tensor/batch/__init__.py
@@ -1,0 +1,4 @@
+"""Validate torch batch data."""
+
+# Copyright (C) 2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
