+++ src/anomalib/dataclasses/validate/tensor/batch/image.py
@@ -1,4 +1,4 @@
-"""Video validation functions."""
+"""Validate torch batch image data."""
 
 # Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
