+++ src/anomalib/dataclasses/validate/tensor/batch/label.py
@@ -1,4 +1,4 @@
-"""Depth validation functions."""
+"""Validate torch batch label data."""
 
 # Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
