+++ src/anomalib/dataclasses/validate/tensor/batch/mask.py
@@ -1,0 +1,4 @@
+"""Validate torch batch mask data."""
+
+# Copyright (C) 2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
