+++ src/anomalib/dataclasses/validate/tensor/batch/score.py
@@ -1,0 +1,4 @@
+"""Validate torch batch score data."""
+
+# Copyright (C) 2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
